added leave category controller
