+++ back-end/app/Http/Controllers/LeaveController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Models\Contracts;
 use App\Models\Leave;
-use App\Models\LeaveCategory;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 
@@ -65,18 +64,5 @@ class LeaveController extends Controller
 
         return response()->json($leaveRequests);
     }
-
-    public function createLeaveCategory(Request $request){
-        $request->validate(
-            ['categoryName' => 'required|string|max:255',
-            'isPaidLeave' => 'required|boolean']
-        );
-
-         LeaveCategory::create([
-            'name' => $request['categoryName'],
-            'isPaidLeave' => $request['isPaidLeave'],
-        ]);
-
-         return response()->json(['message' => 'Leave category created'], 200);
-    }
+    
 }
