+++ back-end/routes/api.php
@@ -2,12 +2,12 @@
 
 use App\Http\Controllers\ContractController;
 use App\Http\Controllers\HealthController;
+use App\Http\Controllers\LeaveCategoryController;
 use App\Http\Controllers\LeaveController;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\RegistrationController;
 use App\Http\Controllers\RolesController;
 use Illuminate\Support\Facades\Route;
-
 
 Route::get('/health', [HealthController::class, 'index']);
 
@@ -32,12 +32,13 @@ Route::prefix('/leave')->group(function (): void {
 
     Route::get('/leave-hours', [LeaveController::class, 'getLeaveHours']);
 
+    Route::prefix('/category')->group(function (): void {
+        Route::post('/', [LeaveCategoryController::class, 'createLeaveCategory']);
+        Route::get('/', [LeaveCategoryController::class, 'getLeaveCategories']);
+    });
 });
 
-Route::prefix('leave-category')->group(function (): void {
-    Route::post('/create', [LeaveController::class, 'createLeaveCategory']);
 
-});
 
 
 Route::prefix('/contract')->group(function () {
